CycleGAN for Desmoking & Dehazing
The CycleGAN method is a deep learning-based approach for unpaired image-to-image translation, which can be used for video desmoking and dehazing. It is a type of generative adversarial network (GAN) that learns to translate images from one domain to another without requiring corresponding pairs of images. In the context of video desmoking and dehazing, the CycleGAN method can be used to translate hazy and smoky video frames to clear and desmoked frames.